Resolve minor UI issues

+++ src/main/angular/client/src/app/components/resume-modal/resume.component.ts
@@ -94,24 +94,27 @@ export class ResumeOrderModalComponent {
           }
           this.constants = this.coreService.convertObjectToArray(res, 'constants');
           this.constants.forEach((item) => {
-            if (!isArray(item.value)) {
-              this.coreService.removeSlashToString(item, 'value');
-              const startChar = item.value.substring(0, 1);
-              const endChar = item.value.substring(item.value.length - 1);
-              if ((startChar === '"' && endChar === '"') || (startChar === "'" && endChar === "'")) {
-                item.value = item.value.substring(1, item.value.length - 1);
-              }
-            } else {
-
-              item.type = 'list';
-              item.value.forEach((val, index) => {
-                item.value[index] = Object.entries(item.value[index]).map(([k1, v1]) => {
-                  return {name: k1, value: v1};
+            if(item.value) {
+              if (!isArray(item.value)) {
+                this.coreService.removeSlashToString(item, 'value');
+                if(typeof item.value == 'string') {
+                  const startChar = item.value.substring(0, 1);
+                  const endChar = item.value.substring(item.value.length - 1);
+                  if ((startChar === '"' && endChar === '"') || (startChar === "'" && endChar === "'")) {
+                    item.value = item.value.substring(1, item.value.length - 1);
+                  }
+                }
+              } else {
+                item.type = 'list';
+                item.value.forEach((val, index) => {
+                  item.value[index] = Object.entries(item.value[index]).map(([k1, v1]) => {
+                    return {name: k1, value: v1};
+                  });
                 });
-              });
-              item.value.forEach((val) => {
-                this.coreService.removeSlashToString(val, 'value');
-              });
+                item.value.forEach((val) => {
+                  this.coreService.removeSlashToString(val, 'value');
+                });
+              }
             }
           });
           this.positions = res.positions.map((pos) => pos.positionString);
